+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Version 4.3.11 so far... [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
 
 - Improved compendium prefab support. Improved and drag/drop support of items.
+- Fix crashes with NND maneuvers. [#4326](https://github.com/dmdorman/hero6e-foundryvtt/issues/4326)
+- Added support for overriding AOE templates with manual targeting [#4358](https://github.com/dmdorman/hero6e-foundryvtt/issues/4358)
 
 ### Version 4.3.11 20260621